requirements.txt
web3
colorama
pyyaml
python-dotenv